+++ packages/dom-css/tsdown.config.ts
@@ -2,7 +2,9 @@ import { defineConfig } from 'tsdown'
 
 export default defineConfig({
   entry: {
-    index: './src/index.ts'
+    index: './src/index.ts',
+    'jsx-runtime': './src/jsx/runtime.ts',
+    'jsx-dev-runtime': './src/jsx/dev-runtime.ts'
   },
   platform: 'neutral',
   format: ['esm'],